SYNC-1571 Move to exception package
