+++ portlets/sync-admin-portlet/docroot/WEB-INF/src/com/liferay/sync/admin/exception/OAuthPortletUndeployedException.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.sync.admin;
+package com.liferay.sync.admin.exception;
 
 import com.liferay.portal.kernel.exception.PortalException;
